Update Model-View link

+++ FridgeManager/src/view/View.java
@@ -9,12 +9,12 @@ public class View
 		this.frame = new Frame();
 	}
 	
-	public void update_fridge_temperature(int value)
+	public void update_fridge_temperature(double value)
 	{
 		frame.fridge_temperature_panel.update_value(value);
 	}
 		
-	public void update_cooler_temperature(int value)
+	public void update_cooler_temperature(double value)
 	{
 		frame.cooler_temperature_panel.update_value(value);
 	}
